Add files via upload
Option #1:  Creating a User Interface I
Create a user interface that has a top bar that includes a menu. The menu should have four items.

When the user selects the first menu option, then the date and time should be printed in a text box.
When the user selects the second menu option, then the text box contents should be written to a text file named "log.txt."
When the user selects the third menu item then the frame background color changes to random color hue of the color green. The menu option should display the initial random hue each time selected for a single execution of the program. See https://www.w3schools.com/colors/colors_picker.aspLinks to an external site.
When the user selects the fourth menu option then the program exits.